requirements.txt
azure-mgmt-storage==21.0.0
azure-storage-blob==12.17.0
python-dotenv==1.0.0
pymongo==4.4.1
requests==2.31.0